вынес генерацию нового вопроса в отдельный файл
