--- src/util/random.ts
@@ -1,7 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export function getRandomInt(min: number, max: number): number {
-  min = Math.ceil(min);
-  max = Math.floor(max);
-
-  return Math.floor(Math.random() * (max - min)) + min;
-}
